refactor: Update imports for EditorGroupsSettings in multiple files
