+++ src/main/java/krasa/editorGroups/actions/SwitchGroupAction.kt
@@ -16,6 +16,7 @@ import com.intellij.ui.PopupHandler
 import krasa.editorGroups.*
 import krasa.editorGroups.icons.EditorGroupsIcons
 import krasa.editorGroups.model.*
+import krasa.editorGroups.settings.EditorGroupsSettings
 import krasa.editorGroups.support.Notifications.showWarning
 import krasa.editorGroups.support.Splitters
 import krasa.editorGroups.support.getVirtualFileByAbsolutePath
