+++ src/main/java/krasa/editorGroups/actions/ToggleAutoSameNameGroupsAction.kt
@@ -4,7 +4,7 @@ import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.ToggleAction
 import com.intellij.openapi.project.DumbAware
-import krasa.editorGroups.EditorGroupsSettings
+import krasa.editorGroups.settings.EditorGroupsSettings
 
 class ToggleAutoSameNameGroupsAction : ToggleAction(), DumbAware {
   override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
