+++ src/main/java/krasa/editorGroups/settings/EditorGroupProjectStorage.kt
@@ -1,4 +1,4 @@
-package krasa.editorGroups
+package krasa.editorGroups.settings
 
 import com.intellij.openapi.components.PersistentStateComponent
 import com.intellij.openapi.components.Service
